requirements.txt
discord.py==2.4.0
mcrcon==0.7.0
mcstatus==11.1.1
tomli==2.0.1